requirements.txt
requests
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
google-cloud-pubsub
google-cloud-functions
fastapi[all]
pydantic
uvicorn
psycopg2-binary
